backend/requirements.txt
flask
flask-cors
gunicorn
psycopg2-binary
python-dotenv
requests
pandas
numpy
dagshub
mlflow
scikit-learn
xgboost
lightgbm
joblib
python-dateutil
catboost